+++ packages/orq-rc/src/orq_ai_sdk/models/createconversationop.py
@@ -199,7 +199,7 @@ class CreateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01khbaq3zz46k3bjxrrsjfz2a1"
+        "conv_01khbm604gkk2rcsnfvbap8r5b"
     )
     r"""Unique conversation identifier with `conv_` prefix."""