+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHBAQ50978RN32QGY1VYYV4B"
+        "01KHBM60N22BRN2YXP7T1QNDBC"
     )
     r"""The unique identifier of the data source"""